Prevent storing mutable references
Since a few iterables are mutable we should copy the values instead of
storing them by reference. Previously a change to an iterable like list,
set or dict would reflect to all other places this reference was shared
with.
This was probably a non issue since exclude lists are most likely not
shared or even mutated, but it can lead to a hard to debug bug.